Refactor: Organize form classes into separate modules under forms package
