+++ forms/__init__.py
@@ -1,0 +1,5 @@
+from .auth import *
+from .event import *
+from .family import *
+from .member import *
+from .user import *
